Abandoning the slicing test trying to test the whole app
Given my standard multi module layout slicing just doesn't seem to work, or doesn't seem to work without heavy use of autowiring.  I want to do that as sparing as possible.  Incidentally the problem I was hitting much of the time had to do with a missing datasource and I believe the solution to that would have been @AutoConfigureTestDatabase.  In the end I want to replace my end 2 end style testing in a box against mock services so the slicing was a nice to have anyway, maybe I'll revisit it knowing what (I think) I know now.