WIP with finding category
So my end 2 end test isn't nearly as configured as I thought, the entityManager or testEntityManager injected into the ApplicationTest doesn't seem to flush it's state for whatever entityManager that is being used by the entityManager in the persistenceDbClient, which I think makes no sense.  It's something transactiony or entityManagery that is going wrong, I verified there is only 1 persistent unit so they SHOULD be talking to the same actual store but who knows.