WIP RestTemplate config
It appears that I'm close now that I understand how the initialization of the RestTemplate works given MockServers and also when HttpClient is in play.  The HttpClient config was removing the MockHttpRequestFactory that's needed to facilitate the test.  I explicitly configure the RestTemplate now and that problem is gone.  Now there is a deserialziation problem but baby steps.